niels/propensities/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibes_eval"
version = "0.2.0"
description = "A library for visualizing model evaluation results"
readme = "README.md"
authors = [
    {name = "VisEval Contributors"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "pyyaml",
    "python-slugify",
    "python-dotenv",
    "cache_on_disk",
    "tqdm>=4.67.1",
    "openweights==0.8.0",
    "localrouter",
    "openai",
    "verifiers",
    "datasets",
]

[project.urls]
"Homepage" = "https://github.com/nielsrolf/viseval"
"Bug Tracker" = "https://github.com/nielsrolf/viseval/issues"

[tool.setuptools]
packages = ["vibes_eval"]
